requirements-3.txt
absl-py==1.0.0
amqp==2.5.1
apns2==0.5.0
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
azure-storage-blob<12
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
bcrypt==3.1.7
billiard==3.6.1.0
boto==2.49.0
boto3==1.18.2
botocore==1.21.65
caniusepython3==7.1.0
cchardet==2.1.4
celery==4.3.0
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
cryptography==2.7
dateutils==0.6.6
decorator==4.4.0
defusedxml==0.7.1
deprecation==2.0.7
distlib==0.2.9.post0
Django==2.2.5
django-adminplus==0.5
django-anymail==6.1.0
django-appconf==1.0.3
django-bulk-update==2.2.0
django-compressor==2.3
django-constance==2.4.0
django-cors-headers==3.1.0
django-crequest==2018.5.11
django-daterange-filter==1.3.0
django-debug-toolbar==2.0
django-environ==0.8.1
django-extra-fields==1.2.4
django-filter==2.2.0
django-fsm==2.6.1
django-grappelli==2.13.1
django-imagekit==4.0.2
django-ipware==2.1.0
django-location-field==2.1.0
git+https://github.com/fireantology/django-logtailer
django-markdown-app==0.9.6
django-model-utils==3.2.0
django-nested-admin==3.2.4
django-pg-utils==0.1.5
django-picklefield==2.0
django-polymorphic==2.1.2
django-push-notifications==2.0.0
django-redis==4.10.0
django-rest-framework==0.1.0
django-storages==1.7.1
django-timezone-field==3.0
django-watson==1.5.2
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-bulk==0.2.1
djangosaml2==1.0.5
drf-nested-routers==0.91
git+https://github.com/razortheory/drf-secure-token
elementpath==2.2.1
factory-boy==2.12.0
Faker==2.0.1
flake8==4.0.1
geographiclib==1.49
geopy==1.20.0
googlemaps==3.1.1
gpxpy==1.3.5
gunicorn==19.9.0
h2==2.6.2
hiredis==1.1.0
hpack==3.0.0
html2text==2019.8.11
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
importlib-metadata==0.20
importlib-resources==5.1.2
invoke==1.3.0
ipdb==0.13.9
ipython==7.10.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.15.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.1.0
jsonfield==2.0.2
kombu==4.6.4
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.2.0
newrelic==5.0.2.126
numpy==1.17.1
ortools==9.2.9972
packaging==19.1
pandas==0.25.1
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
phonenumbers==8.12.48
pickleshare==0.7.5
pilkit==2.0
Pillow==6.1.0
pillow-heif==0.2.1
pinax-stripe==4.4.0
prompt-toolkit==2.0.9
protobuf==3.19.4
psutil==5.6.3
psycopg2==2.8.3
ptyprocess==0.6.0
py-vapid==1.7.0
pycparser==2.19
pyfcm==1.4.7
Pygments==2.4.2
pyheif==0.4
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pysaml2==6.5.1
pysftp==0.2.9
pyshorteners==0.6.1
python-dateutil==2.8.0
python-monkey-business==1.0.0
pytz==2019.2
pywebpush==1.10.0
rcssmin==1.0.6
redis==3.3.8
request-time-tracker==0.0.7
requests==2.22.0
rest-condition==1.0.3
rjsmin==1.1.0
s3transfer==0.5.1
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.3.0
Shapely==1.8.0
six==1.12.0
sqlparse==0.3.0
stripe==2.35.1
tblib==1.7.0
text-unidecode==1.2
threadpoolctl==3.1.0
toml==0.10.2
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.25.11
vine==1.3.0
wcwidth==0.1.7
xmlschema==1.6.0
zipp==0.6.0